Change scalar promotion rules to prefer array types over scalar types.
Currently JAX does not treat Python scalars specially during type promotion. This means that, for example:
1. + np.array([...], np.float32)
ends up as an array of type np.float64. The `1.` is promoted to a default type (here np.float64), and the type promotion of a np.float64 and an np.float32 is an np.float64. This is unlike classic NumPy, which treats scalars specially during type promotion, in particular, preferring the type of an array over the type of a scalar.

This change adds a notion of weak_type to JAX avals. During type promotion, we prefer non-weak types, i.e., the type of the array in the example above, ignoring the type of the scalar.

In contexts where a Python scalar is to be promoted to a NumPy value, a default type is used (e.g., `np.float_`). This change also makes it possible to use 32-bit default types that differ from NumPy's default types. The JAX test suite passes with 32-bit default types. However, we do not yet enable this change or expose it in the API.

More work on Python scalar types.

+++ tests/batching_test.py
@@ -576,7 +576,8 @@ class BatchingTest(jtu.JaxTestCase):
   def testCumProd(self):
    x = np.arange(9).reshape(3, 3) + 1
    y = vmap(lambda x: np.cumprod(x, axis=-1))(x)
-   self.assertAllClose(onp.cumprod(x, axis=1), y, check_dtypes=True)
+   self.assertAllClose(onp.cumprod(x, axis=1, dtype=np.int_), y,
+                       check_dtypes=True)
 
   def testSelect(self):
     pred = onp.array([True, False])
@@ -962,7 +963,8 @@ class BatchingTest(jtu.JaxTestCase):
         key, _ = random.split(key)
         return u, key
 
-      u, _ = lax.while_loop(lambda uk: uk[0] > 0.5, body_fn, (1., key))
+      u, _ = lax.while_loop(lambda uk: uk[0] > 0.5, body_fn,
+                            (np.float64(1.), key))
       return u
 
     print(vmap(f)(random.split(random.PRNGKey(0), 2)))  # no crash
